pyproject.toml
[project]
name = "novel-cassava-disease"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imageio>=2.37.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
